+++ src/core/models/adresse/adresse.js
@@ -8,14 +8,12 @@ var Adresse = (function (_super) {
     __extends(Adresse, _super);
     function Adresse(params) {
         _super.call(this, params);
-        if (!params.nom && !params.cp && params.ville)
-            throw new Error('Le model \'Adresse\' requiert des paramètres \'nom\', \'cp\', \'ville\' non nuls');
-        this.nom = params.nom;
-        this.ligne1 = params.ligne1;
-        this.ligne2 = params.ligne2 || null;
-        this.cp = params.cp;
-        this.ville = params.ville;
-        this.pays = params.pays || null;
+        this.nom = params ? params.nom : null;
+        this.ligne1 = params ? params.ligne1 : null;
+        this.ligne2 = params ? params.ligne2 : null;
+        this.cp = params ? params.cp : null;
+        this.ville = params ? params.ville : null;
+        this.pays = params ? params.pays : null;
     }
     return Adresse;
 })(base_1.default.BaseModel);