+++ src/core/models/identification/identification.ts
@@ -1,20 +1,30 @@
 import Base from '../../../base/base';
 
-export class Identification extends Base.BaseModel {
+export interface IIdentification {
+    SIREN?: string;
+    NIC?: string;
+    APE?: string;
+    RCS?: string;
+    RM?: string;
+    SIRET?: string;   
+}
 
-    constructor(params: any) {
+export class Identification extends Base.BaseModel implements IIdentification {
+
+    constructor(params: IIdentification = null) {
 
         super(params);
 
-        this.SIREN = params.SIREN;
-        this.NIC = params.NIC;
-        this.APE = params.APE;
-        this.RCS = params.RCS || null;
-        this.RM = params.RM || null;
+        this.SIREN = params ? params.SIREN : null;
+        this.NIC = params ? params.NIC : null;
+        this.APE = params ? params.APE : null;
+        this.RCS = params ? params.RCS : null;
+        this.RM = params ? params.RM : null;
 
-        // ON CONSTRUIT LE SIRET
-        this.SIRET = this.SIREN + this.NIC;
+    }
 
+    public get SIRET(): string {
+        return this.SIREN + this.NIC;
     }
 
     SIREN: string;
@@ -22,6 +32,5 @@ export class Identification extends Base.BaseModel {
     APE: string;
     RCS: string;
     RM: string;
-    SIRET: string;
 
 }