+++ src/core/models/models.ts
@@ -1,12 +1,12 @@
 import { Entreprise, IEntreprise } from './entreprise/entreprise';
-import { Adresse } from './adresse/adresse';
-import { Identification } from './identification/identification';
+import { Adresse, IAdresse } from './adresse/adresse';
+import { Identification, IIdentification } from './identification/identification';
 import { Personne, PersonnePhysique, PersonneMorale } from './personne/personne';
 
 export{ 
     Entreprise, IEntreprise,
-    Adresse,
-    Identification,
+    Adresse, IAdresse,
+    Identification, IIdentification,
     Personne,
     PersonneMorale,
     PersonnePhysique