+++ src/core/models/personne/personne.js
@@ -5,18 +5,20 @@ var __extends = (this && this.__extends) || function (d, b) {
 };
 var base_1 = require('../../../base/base');
 var adresse_1 = require('../adresse/adresse');
+var identification_1 = require('../identification/identification');
 var Personne = (function (_super) {
     __extends(Personne, _super);
     function Personne(params) {
+        if (params === void 0) { params = null; }
         _super.call(this, params);
-        this.adresse = params.adresse;
-        this.isEntreprise = params.isEntreprise || false;
-        this.numeroTVA = params.numeroTVA || null;
-        this.telephone = params.telephone || null;
-        this.email = params.email || null;
-        this.siteInternet = params.siteInternet || null;
-        this.fax = params.fax || null;
-        this.identification = params.identification || null;
+        this.adresse = params ? params.adresse : new adresse_1.Adresse({});
+        this.isEntreprise = params ? params.isEntreprise : false;
+        this.numeroTVA = params ? params.numeroTVA : null;
+        this.telephone = params ? params.telephone : null;
+        this.email = params ? params.email : null;
+        this.siteInternet = params ? params.siteInternet : null;
+        this.fax = params ? params.fax : null;
+        this.identification = params ? params.identification : new identification_1.Identification();
     }
     Personne.instanciatePhysiqueOuMorale = function (newPersonne) {
         if (!newPersonne.denominationSociale) {
@@ -42,11 +44,12 @@ exports.Personne = Personne;
 var PersonnePhysique = (function (_super) {
     __extends(PersonnePhysique, _super);
     function PersonnePhysique(params) {
+        if (params === void 0) { params = null; }
         _super.call(this, params);
-        this.civilite = params.civilite || null;
-        this.nom = params.nom || null;
-        this.prenom = params.prenom || null;
-        this.nomCommercial = params.nomCommercial || null;
+        this.civilite = params ? params.civilite : null;
+        this.nom = params ? params.nom : null;
+        this.prenom = params ? params.prenom : null;
+        this.nomCommercial = params ? params.nomCommercial : null;
     }
     return PersonnePhysique;
 })(Personne);
@@ -54,11 +57,12 @@ exports.PersonnePhysique = PersonnePhysique;
 var PersonneMorale = (function (_super) {
     __extends(PersonneMorale, _super);
     function PersonneMorale(params) {
+        if (params === void 0) { params = null; }
         _super.call(this, params);
-        this.forme = params.forme;
-        this.denominationSociale = params.denominationSociale;
-        this.capitalSocial = params.capitalSocial || null;
-        this.isCapitalVariable = params.isCapitalVariable || false;
+        this.forme = params ? params.forme : null;
+        this.denominationSociale = params ? params.denominationSociale : null;
+        this.capitalSocial = params ? params.capitalSocial : null;
+        this.isCapitalVariable = params ? params.isCapitalVariable : false;
     }
     return PersonneMorale;
 })(Personne);