+++ changelog.md
@@ -1,3 +1,8 @@
+#0.2.5
+* Correction de tout les models du Core
+* MEP tests pour Core.Models.Identification
+* MEP tests pour Core.Models.Personne
+
 #0.2.4
 * Correction de l'instanciation du Core.Models.Entreprise
 * Création de Core.Models.IEntreprise : permettant de gérer les paramèters d'instanciation de Core.Models.Entreprise